layout.html
Trying Task5 where the task was to try a website for a Company or Celebrity. My approach was to input as much as a homepage would require..... starting from  
Header, 
Some Important Stuff(highlights), 
Video Input for Advertisement or Digital Branding,
Products or Things to be Featured,
Tried Carousel but in the form of CSS :) 
Contacting Info Section via E-mail,
Lastly a Footer.